services/agent_cerveau/requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
langchain
langchain-core
langchain-groq
